LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focusgroup/tentative/grid-navigation/colspan-expected.txt

r1c1	r1c2	r1c4
r2c1	r2c2	r2c3	r2c4

FAIL A right arrow press should move the focus to the next column, dealing correctly with colspans. assert_equals: expected Element node <td id="r1c2" tabindex="-1" colspan="2">r1c2</td> but got Element node <td id="r1c1" tabindex="0">r1c1</td>
FAIL A down arrow press should move the focus to the right cell on the next row, dealing correctly with colspans. assert_equals: expected Element node <td id="r2c1" tabindex="-1">r2c1</td> but got Element node <td id="r1c1" tabindex="0">r1c1</td>
FAIL A left arrow press should move to the previous column, dealing correctly with the colspans. assert_equals: expected Element node <td id="r1c2" tabindex="-1" colspan="2">r1c2</td> but got Element node <td id="r1c4" tabindex="-1">r1c4</td>
FAIL An up arrow press should move the focus to the right cell on the previous row, dealing correctly with colspans. assert_equals: expected Element node <td id="r1c1" tabindex="0">r1c1</td> but got Element node <td id="r2c1" tabindex="-1">r2c1</td>
